# docs/docs-project/assets/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gateyes Gateway API",
    "description": "LLM API Gateway - OpenAI/Anthropic 兼容\n\n认证方式：\n```\nAuthorization: Bearer <api_key>:<api_secret>\n```\n\n环境变量参考 `configs/config.yaml`。",
    "version": "1.0.0",
    "contact": {
      "name": "Gateyes"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8028",
      "description": "本地"
    }
  ],
  "paths": {
...
    "/admin/providers": {
      "get": {
        "tags": [
          "管理"
        ],
        "summary": "列出所有 Provider",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Provider 列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "vendor": {
                            "type": "string"
                          },
                          "base_url": {
                            "type": "string"
                          },
                          "model": {
                            "type": "string"
                          },
                          "enabled": {
                            "type": "boolean"
                          },
                          "weight": {
                            "type": "integer"
                          },
                          "status": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "管理"
        ],
        "summary": "创建 Provider",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "type"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "my-provider"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "openai",
                      "anthropic",
                      "grpc"
                    ],
                    "example": "openai"
                  },
                  "vendor": {
                    "type": "string",
                    "example": "openai"
                  },
                  "baseURL": {
                    "type": "string",
                    "example": "https://api.example.com/v1"
                  },
                  "apiKey": {
                    "type": "string",
                    "example": "sk-xxx"
                  },
                  "model": {
                    "type": "string",
                    "example": "gpt-4"
                  },
                  "enabled": {
                    "type": "boolean",
                    "example": true
                  },
                  "weight": {
                    "type": "integer",
                    "example": 10
                  }
                }
              },
              "example": {
                "name": "my-provider",
                "type": "openai",
                "vendor": "openai",
                "baseURL": "https://api.example.com/v1",
                "apiKey": "sk-xxx",
                "model": "gpt-4",
                "enabled": true,
                "weight": 10
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider 已创建"
          },
          "400": {
            "description": "参数错误"
          }
        }
      }
    },
    "/admin/providers/{name}": {
      "get": {
        "tags": [
          "管理"
        ],
        "summary": "获取单个 Provider 详情",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider 详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "vendor": {
                      "type": "string"
                    },
                    "base_url": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "enabled": {
                      "type": "boolean"
                    },
                    "weight": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    },
                    "latency_ms": {
                      "type": "number"
                    },
                    "success_rate": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Provider 不存在"
          }
        }
      },
      "put": {
        "tags": [
          "管理"
        ],
        "summary": "更新 Provider",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "baseURL": {
                    "type": "string",
                    "example": "https://api.example.com/v1"
                  },
                  "apiKey": {
                    "type": "string",
                    "example": "sk-xxx"
                  },
                  "model": {
                    "type": "string",
                    "example": "gpt-4"
                  },
                  "enabled": {
                    "type": "boolean",
                    "example": true
                  },
                  "weight": {
                    "type": "integer",
                    "example": 10
                  }
                }
              },
              "example": {
                "baseURL": "https://api.example.com/v1",
                "apiKey": "sk-xxx",
                "model": "gpt-4",
                "enabled": true,
                "weight": 10
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已更新"
          },
          "404": {
            "description": "Provider 不存在"
          }
        }
      },
      "delete": {
        "tags": [
          "管理"
        ],
        "summary": "删除 Provider",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "已删除"
          },
          "404": {
            "description": "Provider 不存在"
          }
        }
      }
    },
...
    "/admin/tenants/{id}/providers": {
      "post": {
        "tags": [
          "租户"
        ],
        "summary": "替换租户 Provider 列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "providers": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "longcat-primary",
                      "minimax"
                    ]
                  }
                }
              },
              "example": {
                "providers": [
                  "longcat-openai",
                  "bigmodel-anthropic"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已更新"
          },
          "400": {
            "description": "Provider 不存在"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "格式：Bearer <api_key>:<api_secret>"
      }
    }
  }
}